tease out stat printing
