+++ test/src/test_main.cpp
@@ -20,6 +20,7 @@
 #include "nakama-cpp/NUtils.h"
 #include "nakama-cpp/NPlatformParams.h"
 #include "nakama-cpp/ClientFactory.h"
+#include "globals.h"
 
 #if defined(__ANDROID__)
 #include <jni.h>
@@ -96,7 +97,19 @@ void NCppTest::tick()
     TaskExecutor::instance().tick();
 }
 
-// *************************************************************
+ostream& printPercent(ostream& os, uint32_t totalCount, uint32_t count)
+{
+    if (totalCount > 0)
+    {
+        os << count * 100 / totalCount << "%";
+    }
+    else
+    {
+        os << "0%";
+    }
+
+    return os;
+}
 
 int runAllTests()
 {
@@ -113,7 +126,13 @@ int runAllTests()
     test_listMatches();
 
     // total stats
-    printTotalStats();
+    uint32_t testsPassed = (g_runTestsCount - g_failedTestsCount);
+
+    NLOG_INFO("Total tests : " + std::to_string(g_runTestsCount));
+    NLOG_INFO("Tests passed: " + std::to_string(testsPassed) +" (");
+    printPercent(cout, g_runTestsCount, testsPassed);
+    NLOG_INFO("Tests failed: " + std::to_string(g_failedTestsCount) + " (");
+    printPercent(cout, g_runTestsCount, g_failedTestsCount);
 
     return getFailedCount() == 0 ? 0 : -1;
 }
